Move vscode helper functions to thier own file
